+++ kotlin-stdlib/commonMain/kotlin/Collections.kt
@@ -2,14 +2,39 @@
 
 package com.javiersc.kotlin.stdlib
 
+/**
+ * Returns second element.
+ *
+ * Throws: `NoSuchElementException` - if the list size is 1.
+ */
 public inline fun <T> Iterable<T>.second(): T = getIndex(2)
 
+/**
+ * Returns third element.
+ *
+ * Throws: `NoSuchElementException` - if the list size is 2.
+ */
 public inline fun <T> Iterable<T>.third(): T = getIndex(3)
 
+/**
+ * Returns forth element.
+ *
+ * Throws: `NoSuchElementException` - if the list size is 3.
+ */
 public inline fun <T> Iterable<T>.forth(): T = getIndex(4)
 
+/**
+ * Returns fifth element.
+ *
+ * Throws: `NoSuchElementException` - if the list size is 4.
+ */
 public inline fun <T> Iterable<T>.fifth(): T = getIndex(5)
 
+/**
+ * Returns penultimate element.
+ *
+ * Throws: `NoSuchElementException` - if the list size is 1.
+ */
 public inline fun <T> Iterable<T>.penultimate(): T {
     return when (this) {
         is List -> this[size - 2]
@@ -30,7 +55,8 @@ public inline fun <T> Iterable<T>.penultimate(): T {
     }
 }
 
-public inline fun <T> Iterable<T>.getIndex(index: Int): T {
+@PublishedApi
+internal inline fun <T> Iterable<T>.getIndex(index: Int): T {
     return when (this) {
         is List -> this[index - 1]
         else -> {
@@ -48,5 +74,5 @@ public inline fun <T> Iterable<T>.getIndex(index: Int): T {
 }
 
 @PublishedApi
-internal fun throwNoSuchElementException(size: Int): Unit =
+internal inline fun throwNoSuchElementException(size: Int): Unit =
     throw NoSuchElementException("Collection size is lower than $size.")